+++ dds/src/dds_async/domain_participant.rs
@@ -9,7 +9,7 @@ use crate::{
             status_condition_actor::StatusConditionActor,
             subscriber_actor::SubscriberActor,
         },
-        utils::actor::ActorAddress,
+        actor::ActorAddress,
     },
     infrastructure::{
         error::{DdsError, DdsResult},
