+++ dds/src/implementation/actors/data_reader_listener_actor.rs
@@ -2,7 +2,7 @@ use dust_dds_derive::actor_interface;
 
 use crate::{
     dds_async::{subscriber::SubscriberAsync, topic::TopicAsync},
-    implementation::utils::actor::ActorAddress,
+    implementation::actor::ActorAddress,
     infrastructure::status::{
         LivelinessChangedStatus, RequestedDeadlineMissedStatus, RequestedIncompatibleQosStatus,
         SampleLostStatus, SampleRejectedStatus, SubscriptionMatchedStatus,
