+++ dds/src/implementation/actors/data_writer_actor.rs
@@ -40,7 +40,7 @@ use crate::{
             },
         },
         udp::udp_transport::UdpTransportWrite,
-        utils::actor::{Actor, ActorAddress},
+        actor::{Actor, ActorAddress},
     },
     infrastructure::{
         self,
