+++ dds/src/implementation/actors/domain_participant_factory_actor.rs
@@ -35,7 +35,7 @@ use crate::{
             writer::RtpsWriter,
         },
         udp::udp_transport::UdpTransportWrite,
-        utils::actor::{Actor, ActorAddress},
+        actor::{Actor, ActorAddress},
     },
     infrastructure::{
         error::{DdsError, DdsResult},
