+++ dds/src/implementation/actors/publisher_actor.rs
@@ -18,7 +18,7 @@ use crate::{
             }, writer::RtpsWriter
         },
         udp::udp_transport::UdpTransportWrite,
-        utils::actor::{Actor, ActorAddress},
+        actor::{Actor, ActorAddress},
     },
     infrastructure::{
         error::{DdsError, DdsResult},
