+++ dds/src/implementation/actors/subscriber_actor.rs
@@ -33,7 +33,7 @@ use crate::{
             },
         },
         udp::udp_transport::UdpTransportWrite,
-        utils::actor::{Actor, ActorAddress},
+        actor::{Actor, ActorAddress},
     },
     infrastructure::{
         error::{DdsError, DdsResult},
