+++ dds/src/implementation/mod.rs
@@ -1,7 +1,7 @@
+pub mod actor;
 pub mod actors;
 pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
 pub mod payload_serializer_deserializer;
 pub mod rtps;
 pub mod udp;
-pub mod utils;
